* W32: "uint" is not available when using the microsoft compiler.    I changed it to uint32_t. Perhaps a better fix is to add "uint"    to "mtypes.h"?!
git-svn-id: svn://svn.minisip.org/minisip/trunk@1726 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/headers/SipHeader.cxx
@@ -177,8 +177,8 @@ MRef<SipHeader *> SipHeader::parseHeader(const string &line){
 				value_zero = value_params[0];
 // 				cerr<<"valueline.substr(2): "+valueline.substr(2)<<endl;
 			} else if( headerType == "Contact" ) {
-				uint ltPos = values[i].find( '<' );
-				uint gtPos = values[i].find( '>' );
+				uint32_t ltPos = (uint32_t) values[i].find( '<' );
+				uint32_t gtPos = (uint32_t) values[i].find( '>' );
 				if( ltPos!=string::npos && gtPos!=string::npos ) {
 					//if the string contains '<' and '>' ...
 					//remove them
